Refactor upload logic into class Conn
